ReporteServicios + 5 a.m culpa de Gustavo

+++ pe.edu.pucp.ferretin.viewmodel/MVentas/MV_ReportesViewModel.cs
@@ -260,7 +260,7 @@ namespace pe.edu.pucp.ferretin.viewmodel.MVentas
             Cliente buscado = null;
             try
             {
-                buscado = MV_ClienteService.obtenerClienteByNroDoc(Convert.ToInt32(searchCliente));
+                buscado = MV_ClienteService.obtenerClienteByNroDoc(Convert.ToInt64(searchCliente));
                 nombreCliente = buscado.nombreCompleto;
                 
             }
